+++ protocols/request-response/CHANGELOG.md
@@ -6,6 +6,9 @@
 - feat: add `Behaviour::send_request_with_addresses()`
   See [PR 5938](https://github.com/libp2p/rust-libp2p/issues/5938).
 
+- fix: don't fail outbound request on `DialError::DialPeerConditionFalse`.
+  See [PR 6000](https://github.com/libp2p/rust-libp2p/pull/6000)
+
 ## 0.28.0
 
 - Deprecate `void` crate.